Update security.py
Change method for getting architecture.
With Python 2 the call to arch will always return x86_64 even on ARM hardware, so the catch to bail on ARM fails.  The value being returned when Intel hardware is expected is 'command' which is part of the error returned due to the command attempted not existing on ARM hardware.