Update security.py
After discussing with jc0b on Slack he pointed out that the main branch of Munkireport doesn't have the fix for detecting the arch on ARM yet, so in stead of relying on that code I've changed it here to use the same method as the updated report common, but standalone so it is not dependent on it. This way the module can be used with any branch of MunkiReport's code base.